cs_backend/requirements.txt
amqp==5.1.1
annotated-types==0.6.0
anyio==4.0.0
asgiref==3.4.1
billiard==4.1.0
boto3==1.21.42
botocore==1.24.42
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
DateTime==5.2
DAWG-Python==0.7.2
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20230430
distro==1.8.0
Django==3.2.5
django-bootstrap-breadcrumbs==0.9.2
django-cors-headers==4.3.0
django-crispy-forms==1.12.0
django-endless-pagination==2.0
django-extensions==3.1.3
django-import-export==2.7.1
django-jazzmin==2.6.0
django-widget-tweaks==1.4.12
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
docopt==0.6.2
drf-yasg==1.21.7
elasticsearch==7.8.1
et-xmlfile==1.1.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.2
MarkupPy==1.14
MarkupSafe==2.1.3
netaddr==0.9.0
numpy
odfpy==1.4.1
openpyxl==3.1.2
orbit-predictor==1.15.0
packaging==23.2
pandas==2.1.1
phonenumbers==8.12.46
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
pydantic==1.10.0
pydantic_core==2.10.1
PyJWT==2.8.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-environ==0.4.54
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.5.2
sgp4==2.22
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.1
tablib==3.5.0
telebot==0.0.5
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
uWSGI==2.0.22
vine==5.0.0
wcwidth==0.2.8
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
yookassa==2.3.6
zope.interface==6.1
